+++ packages/sampler/stories/moonstone-stories/ToggleButton.js
@@ -24,13 +24,13 @@ storiesOf('Moonstone', module)
 		})(() => (
 			<ToggleButton
 				aria-label="toggle button"
-				casing={select('casing', ['preserve', 'sentence', 'word', 'upper'], 'upper')}
-				onClick={action('onClick')}
 				backgroundOpacity={nullify(select('backgroundOpacity', prop.backgroundOpacity))}
+				casing={select('casing', ['preserve', 'sentence', 'word', 'upper'], 'upper')}
 				disabled={boolean('disabled', false)}
+				onToggle={action('onToggle')}
 				small={nullify(boolean('small', false))}
-				toggleOnLabel={text('toggleOnLabel', 'On')}
 				toggleOffLabel={text('toggleOffLabel', 'Off')}
+				toggleOnLabel={text('toggleOnLabel', 'On')}
 			>
 				Missing Toggle Label
 			</ToggleButton>