+++ packages/ui/CHANGELOG.md
@@ -7,6 +7,15 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Changed
 
 - `ui/Toggleable` to forward events on `acitvate` and `deactivate` instead of firing toggled payload. Use `toggle` to handle toggled payload from the event
+## [Unreleased]
+
+### Added
+
+- `ui/Touchable` support to fire `onTap` when a `click` event occurs
+
+### Changed
+
+- `ui/Touchable` custom events `onDown`, `onUp`, `onMove`, and `onTap` to use the event name as the `type` rather than the shorter name (e.g. `onTap` rather than `tap`)
 
 ## [2.0.0-beta.2] - 2018-05-07